+++ callisto_core/tests/test_base.py
@@ -97,9 +97,14 @@ class ReportPostHelper(object):
             'report_update',
             kwargs={'uuid': self.report.uuid, 'step': '0'},
         )
-        data = {'question_3': 'blanket ipsum pillowfight'}
-        response = self.client.post(url, data, follow=True)
+        self.data = {'question_3': 'blanket ipsum pillowfight'}
+        response = self.client.post(url, self.data, follow=True)
         self.assertIn(response.status_code, self.valid_statuses)
+        self.report.refresh_from_db()
+        self.assertEqual(
+            self.decrypted_report['data']['question_3'],
+            self.data['question_3'],
+        )
         return response
 
     def client_post_report_access(self, url):
@@ -204,6 +209,10 @@ class ReportFlowHelper(
         'wizard_builder_data',
         'callisto_core_notification_data',
     ]
+
+    @property
+    def decrypted_report(self):
+        return self.report.decrypted_report(self.passphrase)
 
     def setUp(self):
         self._setup_sites()