+++ callisto_core/utils/settings.py
@@ -46,7 +46,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'nested_admin',
     'widget_tweaks',
-    'wizard_builder',
+    'callisto_core.wizard_builder',
     'callisto_core.delivery',
     'callisto_core.evaluation',
     'callisto_core.notification',
@@ -54,12 +54,11 @@ INSTALLED_APPS = [
     'callisto_core.utils',
 ]
 
-MIDDLEWARE_CLASSES = [
+MIDDLEWARE = [
     'django.contrib.sessions.middleware.SessionMiddleware',
     'django.middleware.common.CommonMiddleware',
     'django.middleware.csrf.CsrfViewMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
-    'django.contrib.auth.middleware.SessionAuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
     'django.middleware.security.SecurityMiddleware',