+++ callisto_core/wizard_builder/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+# TODO: smell this app config
+class WizardBuilderConfig(AppConfig):
+    name = 'callisto_core.wizard_builder'
+    verbose_name = "Wizard Builder"